third_party/blink/web_tests/external/wpt/css/css-viewport/zoom/svg-computed-style-expected.txt

This is a testharness.js-based test.
Found 12 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Property box-shadow value 'inherit' zoom: 2
  assert_equals: expected "rgb(0, 0, 0) 5px 5px 5px 0px" but got "rgb(0, 0, 0) 2.5px 2.5px 2.5px 0px"
[FAIL] Property filter value 'inherit' zoom: 2
  assert_equals: expected "drop-shadow(rgb(0, 0, 0) 5px 5px 5px)" but got "drop-shadow(rgb(0, 0, 0) 2.5px 2.5px 2.5px)"
[FAIL] Property margin value 'inherit' zoom: 2
  assert_equals: expected "-10px 20px 30px -40px" but got "-5px 10px 15px -20px"
[FAIL] Property padding value 'inherit' zoom: 2
  assert_equals: expected "10px 20px 30px 40px" but got "5px 10px 15px 20px"
[FAIL] Property text-decoration-thickness value '4px' no zoom
  assert_equals: expected "4px" but got "auto"
[FAIL] Property text-decoration-thickness value 'inherit' no zoom
  assert_equals: expected "2px" but got "auto"
[FAIL] Property text-decoration-thickness value '4px' zoom: 2
  assert_equals: expected "4px" but got "auto"
[FAIL] Property text-decoration-thickness value 'inherit' zoom: 2
  assert_equals: expected "2px" but got "auto"
[FAIL] Property text-underline-offset value '4px' no zoom
  assert_equals: expected "4px" but got "auto"
[FAIL] Property text-underline-offset value 'inherit' no zoom
  assert_equals: expected "2px" but got "auto"
[FAIL] Property text-underline-offset value '4px' zoom: 2
  assert_equals: expected "4px" but got "auto"
[FAIL] Property text-underline-offset value 'inherit' zoom: 2
  assert_equals: expected "2px" but got "auto"
Harness: the test ran to completion.
